requirements.txt
Django~=3.2.6

# Celery
celery==5.0.5
celery[redis]==5.0.5
django-celery-beat==2.2.0
redis==3.5.3

# Deployment
gunicorn==20.1.0
uvicorn==0.14.0
whitenoise==5.2.0
django-environ==0.4.5
mozilla-django-oidc==1.2.4
psycopg2-binary==2.9.1

# Kubernetes
kubernetes==18.20.0

# IoTDB
apache-iotdb-nightly

# Crispy Forms
django-crispy-forms